split: gen: cros: fdt_support: Use uintptr_t for 32/64-bit compatibility
This fixes problems in these files when building on 64-bit machines.

BUG=chromium-os:19353
TEST=build for Seaboard

Reviewed-on: http://gerrit.chromium.org/gerrit/7644
Tested-by: Simon Glass <sjg@chromium.org>

+++ cros/cmd/cmd_vbexport_test.c
@@ -67,7 +67,7 @@ static int do_vbexport_test_malloc_size(uint32_t size)
 	char *mem = VbExMalloc(size);
 	int i;
 	VbExDebug("Trying to malloc a memory block for %lu bytes...", size);
-	if ((uint32_t)mem % CACHE_LINE_SIZE != 0) {
+	if ((uintptr_t)mem % CACHE_LINE_SIZE != 0) {
 		VbExDebug("\nMemory not algined with a cache line!\n");
 		VbExFree(mem);
 		return 1;
